Pulling test fundemental_zone into own file
